:hammer:[rename]: CommunityDetail->PostDetail 파일 이름 수정
