+++ src/pages/community/PostDetail.tsx
