+++ src/pages/community/index.ts
@@ -1,4 +1,4 @@
 export { default as Community } from "./Community";
-export { default as CommunityDetail } from "./CommunityDetail";
+export { default as PostDetail } from "./PostDetail";
 export { default as SelectDiary } from "./SelectDiary";
 export { default as PostEdit } from "./PostEdit";
